Model refactor

+++ Model/Product.cs
@@ -2,8 +2,8 @@ public class Product {
     public int ProductId { get; set; }
     public string? Name { get; set; }
     public decimal UnitPrice { get; set; }
-    public short UnitsInStock { get; set; }
-    public short QuantityPerUnit { get; set; }
+    public int UnitsInStock { get; set; }
+    public int QuantityPerUnit { get; set; }
